fix: Change lat and long from Long to Double

+++ src/main/java/com/elkusnandi/generalnote/request/ShuttleOutletLocationRequest.java
@@ -10,10 +10,10 @@ public class ShuttleOutletLocationRequest {
     private String name;
 
     @NotBlank
-    private Long latitude;
+    private Double latitude;
 
     @NotBlank
-    private Long longitude;
+    private Double longitude;
 
     @NotBlank
     private String provinceId;
